Finished the Updated Section and Created the Popular section
I have finished the Updated Section of the website and then I have created the popular section of the website.

My plan is to add an API to the sections to allow them to automatically update when i change an entry.

My plan is to add the books section for the website and link it to the landing page using React-Router-Dom to connect them.